Add support for positive numbers less than 1000
Added support to tell if numbers less than 1000 are even or odd. The library's main method returns true if the number is even, and false if otherwise. It is built on top of a *highly optimized* and *robust* algorithm that uses advanced techniques to determine if the number is even.